Simplify in progress state in Analytics activation CTA.
